move concat_imgs
