+++ nilearn/_utils/__init__.py
@@ -21,7 +21,6 @@ from .niimg_conversions import (
     check_niimg,
     check_niimg_3d,
     check_niimg_4d,
-    concat_niimgs,
 )
 from .numpy_conversions import as_ndarray
 
@@ -214,7 +213,6 @@ __all__ = [
     "check_niimg_4d",
     "compare_version",
     "compose_err_msg",
-    "concat_niimgs",
     "copy_img",
     "fill_doc",
     "load_niimg",
