+++ nilearn/image/__init__.py
@@ -3,11 +3,11 @@
 Like, for example, a (3+)D block of data, and an affine.
 """
 from .._utils.niimg import copy_img
-from .._utils.niimg_conversions import concat_niimgs as concat_imgs
 from .image import (
     binarize_img,
     clean_img,
     crop_img,
+    concat_imgs,
     get_data,
     high_variance_confounds,
     index_img,
